divide GetMetricStatistics related code
